+++ semcore/d3-chart/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 - `minRadius` property to Venn.Chart.
 - `index` to children render function of `CompactHorizontalBar` subcomponents.
+- Gaps between bars in StackBar chart.
 
 ## [3.50.0] - 2024-06-28